python/requirements.txt
beautifulsoup4==4.12.2
markdownify==0.11.6
python-dotenv==1.0.0
six==1.16.0
soupsieve==2.5
